+++ lib/data_channel.dart
@@ -1,5 +1,1 @@
-///
-/// data_channel
-///
-
 export 'package:data_channel/src/main.dart';